Refactor files, core was getting really noisy
