+++ core/src/main/scala/OpticLang.scala
@@ -1,8 +1,8 @@
-package org.hablapps.statelesser
+package statelesser
 
 import scalaz._
 
-trait OpticAlg[Expr[_]] {
+trait OpticLang[Expr[_]] {
 
   def flVert[S, A, B](
     l: Expr[Fold[S, A]], 
@@ -54,7 +54,7 @@ trait OpticAlg[Expr[_]] {
   def fl1AsFl[S, A](fl1: Expr[Fold1[S, A]]): Expr[Fold[S, A]]
 }
 
-object OpticAlg {
+object OpticLang {
 
   type TypeNme = String
   type OpticNme = String
@@ -69,40 +69,6 @@ object OpticAlg {
       select: List[Tree] = List(),
       where: Set[Tree] = Set()) {
     
-    import Sql._
-
-    def toSql: SSelect = SSelect(selToSql, tabToSql, whrToSql)
-
-    private def selToSql: SqlSelect = select match {
-      case Var(e) :: Nil => SAll(e)
-      case xs => SList(xs.map(e => SField(treeToExpr(e), "")))
-    }
-
-    private def tabToSql: SqlFrom = SFrom(List(table.toList match {
-      case (Var(nme), GLabel(info)) :: xs => 
-        STable(info.src.nme, nme, xs.map(joinToSql))
-      case _ => throw new Error(s"No table was selected for FROM clause")
-    }))
-
-    private def joinToSql(vt: (OpticAlg.Var, Tree)): SqlJoin = vt match {
-      case (Var(n1), VL(Var(n2), GLabel(inf))) => SEqJoin(
-        s"Dummy_${inf.tgt.nme}", n1, SOn(SProj(n2, inf.nme), SProj(n1, "Dummy_Id")))
-      case _ => throw new Error(s"Don't know how to generate join for '$vt'")
-    }
-
-    private def whrToSql: Option[SqlExp] =
-      where.foldLeft(Option.empty[SqlExp]) {
-        case (None, t) => Some(treeToExpr(t))
-        case (Some(e), t) => Some(SBinOp("AND", e, treeToExpr(t)))
-      }
-
-    private def treeToExpr(t: Tree): SqlExp = t match {
-      case VL(Var(nme), GLabel(info)) => SProj(nme, info.nme)
-      case Op(op, l, r) => SBinOp(op, treeToExpr(l), treeToExpr(r))
-      case Unary(t, op) => SUnOp(op, treeToExpr(t))
-      case _ => throw new Error(s"Don't know how to translate '$t' into SQL")
-    }
-
     def hCompose(other: Semantic): Semantic = {
       val Semantic(t1, s1, w1) = this
       val Semantic(t2, s2, w2) = other
@@ -172,7 +138,7 @@ object OpticAlg {
   case class Unapplied(f: Tree => Tree => Tree) extends Tree
   case class Pred(op: String) extends Tree
 
-  implicit val semantic = new OpticAlg[Const[Semantic, ?]] {
+  implicit val semantic = new OpticLang[Const[Semantic, ?]] {
 
     def flVert[S, A, B](
         l: Const[Semantic, Fold[S, A]], 
@@ -247,19 +213,19 @@ object OpticAlg {
 
     implicit class FoldOps[Expr[_], S, A](
         l: Expr[Fold[S, A]])(implicit 
-        O: OpticAlg[Expr]) {
+        O: OpticLang[Expr]) {
       def >[B](r: Expr[Fold[A, B]]): Expr[Fold[S, B]] = O.flVert(l, r)
     }
 
     implicit class Fold1Ops[Expr[_], S, A](
         l: Expr[Fold1[S, A]])(implicit
-        O: OpticAlg[Expr]) {
+        O: OpticLang[Expr]) {
       def asFold: Expr[Fold[S, A]] = O.fl1AsFl(l)
     }
 
     implicit class AffineFoldOps[Expr[_], S, A](
         l: Expr[AffineFold[S, A]])(implicit
-        O: OpticAlg[Expr]) {
+        O: OpticLang[Expr]) {
       def asFold: Expr[Fold[S, A]] = O.aflAsFl(l)
       def >[B](r: Expr[AffineFold[A, B]]): Expr[AffineFold[S, B]] = 
         O.aflVert(l, r)
@@ -271,7 +237,7 @@ object OpticAlg {
 
     implicit class GetterOps[Expr[_], S, A](
         l: Expr[Getter[S, A]])(implicit
-        O: OpticAlg[Expr]) {
+        O: OpticLang[Expr]) {
       def asFold1: Expr[Fold1[S, A]] = O.gtAsFl1(l)
       def asAffineFold: Expr[AffineFold[S, A]] = O.gtAsAfl(l)
       def >[B](r: Expr[Getter[A, B]]): Expr[Getter[S, B]] = O.gtVert(l, r)
@@ -280,194 +246,11 @@ object OpticAlg {
 
     implicit class GetterIntOps[Expr[_], S](
         l: Expr[Getter[S, Int]])(implicit
-        O: OpticAlg[Expr]) {
+        O: OpticLang[Expr]) {
       def -(r: Expr[Getter[S, Int]]): Expr[Getter[S, Int]] = O.gtSub(l, r)
     }
   }
 
   object syntax extends Syntax
-}
-
-trait CoupleExample[Expr[_]] {
-
-  /* data layer */
-
-  implicit val ev: OpticAlg[Expr]
-  
-  type Couple
-  type Couples = List[Couple]
-  type Person
-  type People = List[Person]
-
-  val couples: Expr[Fold[Couples, Couple]]
-  val her: Expr[Getter[Couple, Person]]
-  val him: Expr[Getter[Couple, Person]]
-  val people: Expr[Fold[People, Person]]
-  val name: Expr[Getter[Person, String]]
-  val age: Expr[Getter[Person, Int]]
-
-  /* logic */
-
-  import ev._
-  import OpticAlg.syntax._
-
-  def getPeople: Expr[People => List[Person]] =
-    getAll(people)
-
-  def getPeopleName: Expr[People => List[String]] =
-    getAll(people > name.asFold1.asFold)
-
-  def getPeopleNameAndAge: Expr[People => List[(String, Int)]] =
-    getAll(people > (name * age).asFold1.asFold)
-
-  def getHerNames: Expr[Couples => List[String]] =
-    getAll(couples > her.asFold1.asFold > name.asFold1.asFold)
-
-  def getHerNameAndAge: Expr[Couples => List[(String, Int)]] =
-    getAll(couples > her.asFold1.asFold > (name * age).asFold1.asFold)
-
-  def getHerNameAndAge2: Expr[Couples => List[(String, Int)]] =
-    getAll(couples > ((her > name) * (her > age)).asFold1.asFold)
-
-  def getHerNameAndAge3: Expr[Couples => List[(String, Int)]] =
-    getAll(couples > (her > name * age).asFold1.asFold)
-
-  def getPeopleGt30: Expr[People => List[(String, Int)]] =
-    getAll(people > (name.asAffineFold * 
-      (age.asAffineFold > filtered (gt(30)))).asFold)
-
-  def getHerGt30: Expr[Couples => List[(String, Int)]] =
-    getAll(couples > her.asFold1.asFold > (name.asAffineFold * 
-      (age.asAffineFold > filtered (gt(30)))).asFold)
-
-  def getHerGt30_2: Expr[Couples => List[(String, Int)]] =
-    getAll(couples > ((her > name).asAffineFold * 
-      ((her > age).asAffineFold > filtered (gt(30)))).asFold)
-
-  def getHerNameGt30: Expr[Couples => List[String]] =
-    getAll(couples > her.asFold1.asFold > (name.asAffineFold <* 
-      (age.asAffineFold > filtered (gt(30)))).asFold)
-  
-  def getHerNameGt30_2: Expr[Couples => List[String]] =
-    getAll(couples > ((her > name).asAffineFold <* 
-      ((her > age).asAffineFold > filtered (gt(30)))).asFold)
-
-  def difference: Expr[Couples => List[(String, Int)]] =
-    getAll(couples > 
-      ((her > name).asAffineFold * 
-        (((her > age) - (him > age)).asAffineFold > filtered (gt(0)))).asFold)
-
-  def differenceName: Expr[Couples => List[String]] =
-    getAll(couples > 
-      ((her > name).asAffineFold <* 
-        (((her > age) - (him > age)).asAffineFold > filtered (gt(0)))).asFold)
-
-  def differenceName2: Expr[Couples => List[String]] =
-    getAll(couples > 
-      ((her > name).asAffineFold * 
-        (((her > age) - (him > age)).asAffineFold > filtered (gt(0))) > 
-          first.asAffineFold).asFold)
-
-  def dummyNameAndAge: Expr[People => List[(String, Int)]] =
-    getAll(people > ((name.asAffineFold * ((name * age > 
-      first * second > 
-      second * first > 
-      second * first > 
-      second).asAffineFold 
-      > filtered (gt(30))
-      > filtered (gt(40))))).asFold)
-}
-
-object CoupleExample {
-  import OpticAlg._
-
-  implicit val semantic = new CoupleExample[Const[Semantic, ?]] {
-
-    implicit val ev = OpticAlg.semantic
-
-    type Couple = Unit
-    type Person = Unit
-
-    val couples = {
-
-      val oi = OpticInfo(
-        "couples", 
-        TypeInfo("Couples", false), 
-        TypeInfo("Couple", true))
-
-      Const(Semantic(Map(Var("c") -> GLabel(oi)), List(Var("c"))))
-    }
-
-    val her = {
-
-      val oi = OpticInfo(
-        "her", 
-        TypeInfo("Couple", false), 
-        TypeInfo("Person", true))
-
-      Const(Semantic(Map(Var("w") -> GLabel(oi)), List(Var("w"))))
-    }
-
-    val him = {
-
-      val oi = OpticInfo(
-        "him", 
-        TypeInfo("Couple", false), 
-        TypeInfo("Person", true))
-
-      Const(Semantic(Map(Var("m") -> GLabel(oi)), List(Var("m"))))
-    }
-
-    val people = {
-
-      val oi = OpticInfo(
-        "people", 
-        TypeInfo("People", false), 
-        TypeInfo("Person", true))
-
-      Const(Semantic(Map(Var("p") -> GLabel(oi)), List(Var("p"))))
-    }
-
-    val name = {
-
-      val oi = OpticInfo(
-        "name", 
-        TypeInfo("Person", false), 
-        TypeInfo("Person", true))
-
-      Const(Semantic(Map(), List(GLabel(oi))))
-    }
-
-    val age = {
-
-      val oi = OpticInfo(
-        "age", 
-        TypeInfo("Person", false), 
-        TypeInfo("Person", true))
-
-      Const(Semantic(Map(), List(GLabel(oi))))
-    }
-  }
-}
-
-object Run extends App {
-  import CoupleExample.semantic._
-
-  println(getPeople.getConst.toSql)
-  println(getPeopleName.getConst.toSql)
-  println(getPeopleNameAndAge.getConst.toSql)
-  println(getHerNames.getConst.toSql)
-  println(getHerNameAndAge.getConst.toSql)
-  println(getHerNameAndAge2.getConst.toSql)
-  println(getHerNameAndAge3.getConst.toSql)
-  println(getPeopleGt30.getConst.toSql)
-  println(getHerGt30.getConst.toSql)
-  println(getHerGt30_2.getConst.toSql)
-  println(getHerNameGt30.getConst.toSql)
-  println(getHerNameGt30_2.getConst.toSql)
-  println(difference.getConst.toSql)
-  println(differenceName.getConst.toSql)
-  println(differenceName2.getConst.toSql)
-  println(dummyNameAndAge.getConst.toSql)
 }
